Create installer artifact in appveyor
Generate a windows installer as an artifact, generated by CPack.

The Wix patch patches the environment with the nvim-qt exe path,
The fragment Id "CM_CP_bin.nvim_qt.exe" was retrieved from the
xml generated by cpack (the dash in the id was automatically replaced
by underscore).

The license file needs to be renamed for the installer to pick it up.
So a copy with .txt extension is created at configuration time.